+++ package_monitor/managers.py
@@ -14,12 +14,12 @@ from .app_settings import (
     PACKAGE_MONITOR_SHOW_ALL_PACKAGES,
     PACKAGE_MONITOR_SHOW_EDITABLE_PACKAGES,
 )
-from .core import (
+from .core.distribution_packages import (
     DistributionPackage,
     compile_package_requirements,
     gather_distribution_packages,
-    update_packages_from_pypi,
 )
+from .core.pypi import update_packages_from_pypi
 
 TERMINAL_MAX_LINE_LENGTH = 4095