+++ package_monitor/tests/factories/factories.py
@@ -4,10 +4,10 @@ import factory
 import factory.fuzzy
 from packaging.requirements import Requirement
 
-from package_monitor.core import DistributionPackage
+from package_monitor.core.distribution_packages import DistributionPackage
 from package_monitor.models import Distribution
 
-from .stubs import ImportlibDistributionStub, Pypi, PypiInfo, PypiRelease, PypiUrl
+from .stubs import MetadataDistributionStub, Pypi, PypiInfo, PypiRelease, PypiUrl
 
 faker = factory.faker.faker.Faker()
 
@@ -35,6 +35,8 @@ class PypiInfoFactory(factory.Factory):
 
 
 class PypiFactory(factory.Factory):
+    """A data object on PyPI. Create from DistributionPackage"""
+
     class Meta:
         model = Pypi
         exclude = ("distribution",)
@@ -58,9 +60,9 @@ class PypiFactory(factory.Factory):
     urls = factory.LazyAttribute(lambda o: [PypiUrlFactory()])
 
 
-class ImportlibDistributionStubFactory(factory.Factory):
+class MetadataDistributionStubFactory(factory.Factory):
     class Meta:
-        model = ImportlibDistributionStub
+        model = MetadataDistributionStub
 
     name = factory.Faker("last_name")
     # files = ["dummy_1/file_1.py", "dummy_1/__init__.py"]